% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birdsEye.r
\docType{package}
\name{birdsEye}
\alias{birdsEye}
\title{birdsEye: Geographic analysis}
\description{
This package contains tools for converting between \code{Spatial*} objects and interpolating between \code{SpatialPolygons*}.
}
\details{
Create an issue on \href{https://github.com/adamlilith/birdsEye/issues}{GitHub}.
}
\author{
Adam B. Smith
}
