% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordsToPoly.r
\name{coordsToPoly}
\alias{coordsToPoly}
\alias{coordsToLine}
\alias{coordsToLines}
\title{Convert coordinates to spatial polygon(s) or spatial line(s)}
\usage{
coordsToPoly(x, crs, closePoly = TRUE, id = 1)

coordsToLine(x, crs, id = 1)

coordsToLines(x, crs, id = seq_along(x))
}
\arguments{
\item{x}{Either a two-column matrix or data frame with coordinates (first column is longitude, second is latitude), or a list of matrices or data frames, or an object from which coordinates can be extracted using \code{link[sp]{coordinates}}.}

\item{crs}{A character string or an object of class \code{\link[sp]{CRS}} which species the coordinate reference system (proj4string) for the polygon. If the object in \code{x} is a spatial object the CRS will be obtained from it unless \code{crs} is specified.}

\item{closePoly}{Logical, if \code{TRUE} (default), then close the polygon.}

\item{id}{Integer or character, ID value for the object(s).}
}
\value{
A spatial polygon object.
}
\description{
This function converts coordinates to a SpatialPolygon object. To close the polygon the last set of coordinates must be equal to the first. If they are not, then the first set is added to the end. Note that it is possible to create an improper polygon (i.e., that crosses itself).
}
\section{Functions}{
\itemize{
\item \code{coordsToLine}: Convert coordinates to a spatial line

\item \code{coordsToLines}: Convert a list of coordinates to spatial lines
}}

\examples{
x <- c(-116.57, -99.13, -62.95, -65.21, -90.08)
y <- c(62.09, 75.50, 73.29, 68.34, 66.43)
xy <- cbind(x, y)
poly <- coordsToPoly(xy, getCRS('wgs84'))
line <- coordsToLine(xy, getCRS('wgs84'))
par(mfrow=c(1, 2))
plot(poly)
plot(line)
}
\seealso{
\code{\link[enmSdm]{SpatialPolygon}}
}
