% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipsoid.r
\name{ellipsoid}
\alias{ellipsoid}
\alias{datum}
\title{Ellipsoid from a Spatial* object, CRS, or proj4 string}
\usage{
ellipsoid(x)

datum(x)
}
\arguments{
\item{x}{SpatialPoints*, SpatialPolygons*, or SpatialLines* object, or an object of class \code{CRS} or \code{character} representing a proj4 string.}
}
\value{
Character.
}
\description{
These functions extract the ellipsoid or datum from a spatial object, a CRS object, or from a proj4 string.
}
\section{Functions}{
\itemize{
\item \code{datum}: Datum from a Spatial* object, CRS, or proj4 string
}}

\examples{
crs <- enmSdm::getCRS('albersNA')
ellipsoid(crs)
datum(crs)
}
