% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomToCoords.r
\name{geomToCoords}
\alias{geomToCoords}
\title{Extract vertex points from spatial polygon or lines object}
\usage{
geomToCoords(x, n = NULL, out = "matrix", removeClose = TRUE)
}
\arguments{
\item{x}{SpatialPolygon, SpatialPolygonDataFrame, SpatiaLines, or SpatialLinesDataFrame.}

\item{n}{\code{NULL} or positive integer(s) indicating which sub-polygon(s) or line(s) to extract. The default \code{NULL} is to extract points for all vertices of all sub-geometries. To see the number of sub-polygons or lines in an object use \code{\link[birdsEye]{countSubGeoms}}.}

\item{out}{Character, either \code{'matrix'} (default) or \code{'pts'}, indicating the returned object should either be a matrix with coordinates (default) or an object of class SpatialPoints}

\item{removeClose}{Logical, if \code{TRUE} (default), then remove the last pair of coordinates (for polygons only--ignored if lines). Most spatial polygons have the same set of first and last coordinates to complete the polygon.}
}
\value{
A two-column matrix or a SpatialPoints object.
}
\description{
This function extracts coordinates of vertices of a polygon or line from a SpatialPolygon, SpatialPolygonDataFrame, SpatiaLines, or SpatialLinesDataFrame.
}
\examples{
data(mad0)
island <- subGeomFromGeom(mad0, 1)
pts <- geomToCoords(island, 1)
plot(island)
points(pts)
}
\seealso{
\code{\link[birdsEye]{countSubGeoms}}, \code{\link[birdsEye]{subGeomFromGeom}}
}
