% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subGeom.r
\name{subGeom}
\alias{subGeom}
\title{Extract a sub-polygon or coordinates thereof from a SpatialPolygons* object}
\usage{
subGeom(x, n)
}
\arguments{
\item{x}{SpatialPolygon or SpatialPolygonDataFrame.}

\item{n}{Positive integer(s) indicating which sub-polygon(s) to obtain. To see the number of sub-polygons in an object use \code{\link[birdsEye]{countSubGeoms}}. If \code{NULL}, then the function basically returns \code{x}, but with all sub-geometries split (i.e., a multi-part polygon/lines).}

\item{out}{Character, either \code{'poly'} (default) or \code{'coords'}, indicating the returned object should either be a spatial polygon or a matrix of coordinates.}
}
\value{
A spatial polygon object or a two-column matrix.
}
\description{
This function extracts a sub-polygon or the coordinates of vertices a sub-polygon from a SpatialPolygon or SpatialPolygonDataFrame object.
}
\examples{
data(mad0)
sub <- subGeom(mad0, 1)
pts <- geomToCoords(mad0, 1)
plot(sub)
points(pts)

N <- countSubGeoms(mad0)
sub <- subGeom(mad0, N)
plot(sub)

par(mfrow=c(4, 4))
for (n in 1:16) {
sub <- subGeom(mad0, n)
plot(sub)
}
}
\seealso{
\code{\link[birdsEye]{countSubGeoms}}, \code{\link[birdsEye]{geomToCoords}}
}
